clean up the services
- now that we don't use the DMCP, can massively simplify
  the services stuff
- also move them to live under the boot package
