+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/boot/impl/ReplacementDuplicationStrategy.java
@@ -1,4 +1,4 @@
-package org.hibernate.reactive.service;
+package org.hibernate.reactive.boot.impl;
 
 import org.hibernate.event.service.spi.DuplicationStrategy;
