+++ hibernate-reactive-core/src/main/resources/META-INF/services/org.hibernate.integrator.spi.Integrator
@@ -1,1 +1,1 @@
-org.hibernate.reactive.service.ReactiveIntegrator
+org.hibernate.reactive.boot.impl.ReactiveIntegrator
